automated-report-generator/requirements.txt
# Automated Report Generator Dependencies
streamlit>=1.31.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
reportlab>=4.0.0
python-docx>=1.1.0
openai>=1.12.0
openpyxl>=3.1.0
pyyaml>=6.0.0
pillow>=10.2.0
